added Header component *notWorking
